requirements.txt
# API & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.7.0
pydantic-settings==2.3.0

# PDF & arXiv
pymupdf==1.24.5
arxiv==2.1.3

# RAG & Embeddings
llama-index-core==0.10.68
llama-index-vector-stores-chroma==0.1.10
chromadb==0.5.5
fastembed==0.7.4
sentence-transformers==3.0.1
rank_bm25==0.2.2

# LLM
groq==1.1.1
openai==2.28.0
google-generativeai>=0.8.0

# NLP & Graph
spacy==3.7.5
networkx==3.3

# Manim
manim==0.18.1

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
